@@ -11,34 +11,34 @@
   },
   "dependencies": {
     "@babel/core": "^7.19.3",
-    "@chakra-ui/icons": "^2.0.11",
-    "@emotion/react": "^11.10.4",
-    "@emotion/styled": "^11.10.4",
     "core-js": "^3.25.5",
     "firebase": "^9.12.1",
     "formik": "^2.2.9",
-    "framer-motion": "^6.5.1",
     "lodash": "^4.17.21",
-    "next": "12.2.3",
+    "next": "13.1.2",
+    "nookies": "^2.5.2",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-firebase-hooks": "^5.0.3",
-    "swr": "^1.3.0"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@chakra-ui/react": "^2.3.5",
-    "@chakra-ui/system": "^2.2.12",
-    "@chakra-ui/theme-tools": "^2.0.12",
+    "@firebase/app-types": "^0.9.0",
+    "@tailwindcss/typography": "^0.5.9",
     "@types/lodash": "^4.14.186",
     "@types/node": "18.6.1",
     "@types/react": "18.0.15",
     "@types/react-dom": "18.0.6",
+    "autoprefixer": "^10.4.13",
+    "encoding": "^0.1.13",
     "eslint": "8.20.0",
-    "eslint-config-next": "12.2.3",
+    "eslint-config-next": "13.1.2",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-react": "^7.31.10",
     "eslint-plugin-unused-imports": "^2.0.0",
+    "firebase-admin": "^11.4.1",
+    "install": "^0.13.0",
     "newsapi": "^2.4.1",
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.2.4",
     "typescript": "4.7.4"
   }
 }
